refactor(frontend): organize agent studio chat composer
