+++ packages/frontend/src/pages/agents/chat-composer/context-usage/use-active-session-context-usage.ts
@@ -19,7 +19,7 @@ type ContextUsageCache = {
   value: NonNullable<AgentStudioContextUsage>;
 };
 
-export const useAgentStudioContextUsage = ({
+export const useActiveSessionContextUsage = ({
   activeSession,
   activeSessionMessages,
   activeSessionLiveContextUsage,
