+++ packages/frontend/src/pages/agents/chat-composer/model-selection/selection-resolution.ts
@@ -1,0 +1,81 @@
+import type { AgentModelCatalog, AgentModelSelection } from "@openducktor/core";
+import { coerceVisibleSelectionToCatalog } from "../../agents-page-selection";
+
+export const resolveRoleDefaultSelectionForComposer = ({
+  hasActiveSession,
+  composerCatalog,
+  isAwaitingRepoSettingsForWorkspaceRepoPath,
+  roleDefaultSelection,
+}: {
+  hasActiveSession: boolean;
+  composerCatalog: AgentModelCatalog | null;
+  isAwaitingRepoSettingsForWorkspaceRepoPath: boolean;
+  roleDefaultSelection: AgentModelSelection | null;
+}): AgentModelSelection | null => {
+  if (hasActiveSession) {
+    return roleDefaultSelection;
+  }
+  if (!composerCatalog) {
+    return isAwaitingRepoSettingsForWorkspaceRepoPath ? null : roleDefaultSelection;
+  }
+  return coerceVisibleSelectionToCatalog(composerCatalog, roleDefaultSelection);
+};
+
+export const resolveSelectionCatalogLoading = ({
+  hasActiveSession,
+  activeSessionIsLoadingModelCatalog,
+  activeSessionModelCatalog,
+  composerCatalog,
+  isLoadingComposerCatalog,
+}: {
+  hasActiveSession: boolean;
+  activeSessionIsLoadingModelCatalog: boolean;
+  activeSessionModelCatalog: AgentModelCatalog | null;
+  composerCatalog: AgentModelCatalog | null;
+  isLoadingComposerCatalog: boolean;
+}): boolean => {
+  if (hasActiveSession) {
+    return activeSessionIsLoadingModelCatalog && !activeSessionModelCatalog && !composerCatalog;
+  }
+  return isLoadingComposerCatalog;
+};
+
+export const resolveSelectedModelSelection = ({
+  activeSessionSelectedModel,
+  draftSelection,
+  roleDefaultSelectionForComposer,
+  fallbackCatalogSelection,
+}: {
+  activeSessionSelectedModel: AgentModelSelection | null;
+  draftSelection: AgentModelSelection | null;
+  roleDefaultSelectionForComposer: AgentModelSelection | null;
+  fallbackCatalogSelection: AgentModelSelection | null;
+}): AgentModelSelection | null => {
+  return (
+    activeSessionSelectedModel ??
+    draftSelection ??
+    roleDefaultSelectionForComposer ??
+    fallbackCatalogSelection ??
+    null
+  );
+};
+
+export const resolveSelectionForNewSession = ({
+  draftSelection,
+  roleDefaultSelectionForComposer,
+  selectionCatalog,
+  fallbackCatalogSelection,
+}: {
+  draftSelection: AgentModelSelection | null;
+  roleDefaultSelectionForComposer: AgentModelSelection | null;
+  selectionCatalog: AgentModelCatalog | null;
+  fallbackCatalogSelection: AgentModelSelection | null;
+}): AgentModelSelection | null => {
+  return (
+    draftSelection ??
+    roleDefaultSelectionForComposer ??
+    coerceVisibleSelectionToCatalog(selectionCatalog, fallbackCatalogSelection) ??
+    fallbackCatalogSelection ??
+    null
+  );
+};
