+++ packages/frontend/src/pages/agents/chat-composer/model-selection/use-active-session-model-selection-repair.ts
@@ -1,9 +1,9 @@
 import type { AgentModelCatalog, AgentModelSelection } from "@openducktor/core";
 import { useEffect } from "react";
-import { isSameSelection } from "../agents-page-selection";
+import { isSameSelection } from "../../agents-page-selection";
 import { resolveActiveSessionModelSelection } from "./model-selection-preferences";
 
-export const useAgentStudioActiveSessionModelRepair = ({
+export const useActiveSessionModelSelectionRepair = ({
   activeExternalSessionId,
   activeSessionModelCatalog,
   activeSessionSelectedModel,
