+++ packages/frontend/src/pages/agents/chat-composer/model-selection/use-draft-model-selection.ts
@@ -1,7 +1,7 @@
 import type { AgentModelCatalog, AgentModelSelection, AgentRole } from "@openducktor/core";
 import { useCallback, useEffect, useRef, useState } from "react";
 import type { RepoSettingsInput } from "@/types/state-slices";
-import { emptyDraftSelections, isSameSelection } from "../agents-page-selection";
+import { emptyDraftSelections, isSameSelection } from "../../agents-page-selection";
 import { resolveDraftModelSelection } from "./model-selection-preferences";
 
 const emptyDraftSelectionTouchedByRole = (): Record<AgentRole, boolean> => ({
